[opsem] Split RawMemManager into header and body
